+++ opentelemetry/src/metrics/mod.rs
@@ -1,6 +1,5 @@
 //! # OpenTelemetry Metrics API
 
-use std::cmp::Ordering;
 use std::hash::{Hash, Hasher};
 use std::result;
 use std::sync::Arc;
@@ -88,19 +87,6 @@ impl Hash for KeyValue {
             Value::I64(i) => i.hash(state),
             Value::String(s) => s.hash(state),
         };
-    }
-}
-
-impl PartialOrd for KeyValue {
-    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
-        Some(self.cmp(other))
-    }
-}
-
-/// Ordering is based on the key only.
-impl Ord for KeyValue {
-    fn cmp(&self, other: &Self) -> Ordering {
-        self.key.cmp(&other.key)
     }
 }
 
@@ -297,27 +283,6 @@ mod tests {
             let kv1 = KeyValue::new("key", random_value);
             let kv2 = KeyValue::new("key", random_value);
             assert_eq!(hash_helper(&kv1), hash_helper(&kv2));
-        }
-    }
-
-    #[test]
-    fn kv_float_order() {
-        // TODO: Extend this test to all value types, not just F64
-        let float_vals = [
-            0.0,
-            1.0,
-            -1.0,
-            f64::INFINITY,
-            f64::NEG_INFINITY,
-            f64::NAN,
-            f64::MIN,
-            f64::MAX,
-        ];
-
-        for v in float_vals {
-            let kv1 = KeyValue::new("a", v);
-            let kv2 = KeyValue::new("b", v);
-            assert!(kv1 < kv2, "Order is solely based on key!");
         }
     }